[FEATURE] Add some simple errorhandling
Add a new UnknownMessageException which is thrown whenever de message
from the queue does not implements the MessageInterface. Then wrap the
QueueWorker in a try/catch to fetch the error and proceed withot failing
completely.
A message that is failed is removed from the queue, so it is not retried.